requirements.txt
aiogram == 3.1.1
python-dotenv == 1.0.0
SQLAlchemy~=2.0.22
psycopg2 == 2.9.9
gino~=1.0.1
asyncpg~=0.28.0